Temporarily removed zip support in view set loader because it did not appear to be working consistently and caused problems for non-zip datasets

+++ Java/ReflectanceAcquisition/src/tetzlaff/halfwayfield/HalfwayFieldRenderer.java
@@ -91,8 +91,14 @@ public class HalfwayFieldRenderer<ContextType extends Context<ContextType>> impl
 	public void cleanup() 
 	{
 		ulfRenderer.cleanup();
-		halfwayField.diffuseTexture.delete();
-		halfwayField.normalTexture.delete();
+		if (halfwayField.diffuseTexture != null)
+		{
+			halfwayField.diffuseTexture.delete();
+		}
+		if (halfwayField.normalTexture != null)
+		{
+			halfwayField.normalTexture.delete();
+		}
 	}
 
 	@Override
